+++ CHANGES.md
@@ -5,7 +5,8 @@ Unreleased
 * atdd: Workaround d compiler bug regarding declaration order when using aliases (#393)
         Algebraic data types (SumType) now uses `alias this` syntax.
 * atdgen: Add support for `<json open_enum>` in Melange (#401)
-* atdcpp: Initial Release
+* atdcpp: Initial Release (#404)
+* atdcpp: Use `double` c++ type as default floating point type (#411)
 
 2.15.0 (2023-10-26)
 -------------------